waste_ai/api/requirements.txt
Flask-CORS
flask
Pillow
ultralytics
gunicorn